Added console output to launched (main) application.
Added a new private method "MessageConsole" for finding (or creating) a
console with a given name.
Code adapted from the following page:
https://wiki.eclipse.org/FAQ_How_do_I_write_to_the_console_from_a_plug-in%3F

Using this new method, a new console named Job-Notifier is created at
the start of the plugin.

Once installed, the plugin will now print each started and finished job
in the Job-Notifier console.